Finished event docs. Refactored the Subject methods. Removed Listener from View.
