+++ confs/composer/event.json
@@ -22,6 +22,9 @@
     "autoload": {
         "psr-0": {
             "Titon\\Event": "src/"
-        }
+        },
+        "files": [
+            "src/Titon/Event/bootstrap.hh"
+        ]
     }
 }
