+++ docs/en/packages/event/annotations.md
@@ -3,3 +3,5 @@
 The following annotations are available in the Event package.
 
 ## Observer ##
+
+Coming soon... once the annotations package is redone.
