+++ src/Titon/Event/Listener.hh
@@ -7,9 +7,6 @@
 
 namespace Titon\Event;
 
-type ListenerMap = Map<string, mixed>;
-type ListenerOption = shape('method' => string, 'priority' => int, 'once' => bool);
-
 /**
  * The Listener interface defines groups of observers to subscribe for events.
  *
