+++ src/Titon/Event/Observer.hh
@@ -11,8 +11,6 @@ use \Closure;
 use \ReflectionFunction;
 use \ReflectionMethod;
 
-type ParamList = array<mixed>;
-
 /**
  * The Observer class represents a single callback in the event stack.
  * It will be executed anytime an event is triggered that the observer is subscribed for.
